+++ tests/Unit/Config/JsonEncodingOptionsTest.php
@@ -30,7 +30,7 @@ class JsonEncodingOptionsTest extends TestCase
 
     public function testCustomHeaders(): void
     {
-        $response = $this->call('GET', '/graphql', [
+        $response = $this->call('POST', '/graphql', [
             'query' => $this->queries['examples'],
         ]);